refactor: 질문 응답 helper 이름 변경
